changes in catkin script, added ROS install script

+++ scripts/launch_catkin.sh
@@ -8,6 +8,7 @@ echo "  0=exit"
 echo "  1=vic_takeoff"
 echo "  2=vic_mission"
 echo "  3=vic_coordinates"
+echo "  4=vic_mission_3drones"
 
 read program
 
@@ -15,6 +16,7 @@ a=0
 b=1
 c=2
 d=3
+e=4
 
 if [ $program -eq $a ]
 then
@@ -48,5 +50,15 @@ then
 	cd rosbag_files
 	rosbag record -a __name:=current_bag > /dev/null 2>&1 &
 	rosrun vic_package vic_coordinates
+	rosnode kill /current_bag > /dev/null 2>&1
+fi
+
+if [ $program -eq $e ]
+then
+	echo "Running vic_mission_3drones"
+	cd
+	cd rosbag_files
+	rosbag record -a __name:=current_bag > /dev/null 2>&1 &
+	rosrun vic_package vic_mission_3drones
 	rosnode kill /current_bag > /dev/null 2>&1
 fi
